// packages/Common/Types/CookieName.ts
enum CookieName {
  UserID = "user-id",
  Email = "user-email",
  Token = "user-token",
  RefreshToken = "user-refresh-token",
  GlobalSSOToken = "global-sso-token",
  Name = "user-name",
  Timezone = "user-timezone",
  IsMasterAdmin = "user-is-master-admin",
  ProfilePicID = "user-profile-pic-id",
  StatusPageMasterPassword = "status-page-master-password",
  DashboardMasterPassword = "dashboard-master-password",
}

export default CookieName;